+++ python/lib/fortran_namelist.py
@@ -184,6 +184,16 @@ class InputNdvi:
         self.variables.update({'&ndvi_io_extpar': {'ndvi_buffer_file'}})
 
 
+class InputEdgar:
+    '''
+    define structure of  namelist "INPUT_edgar"
+    '''
+
+    def __init__(self):
+
+        self.variables = {'&edgar_io_extpar': {'edgar_buffer_file'}}
+
+
 class InputEmiss:
     '''
     define structure of former namelist "INPUT_EMISS"